create prince_handler module
- move structs and constructor/destructor to module
- set up cmake for new module
